Refactoring out into multiple files to enhance readability and findability of methods
