+++ PowerPoint_Adapter/CRUD/Update/_IUpdateSlide.cs
@@ -1,0 +1,86 @@
+/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2024, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using BH.oM.Adapter;
+using BH.oM.PowerPoint;
+using DocumentFormat.OpenXml;
+using Spreadsheet = DocumentFormat.OpenXml.Spreadsheet;
+using DocumentFormat.OpenXml.Presentation;
+using DocumentFormat.OpenXml.Packaging;
+using Drawing = DocumentFormat.OpenXml.Drawing;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.IO;
+using BH.Engine.Base;
+using BH.Engine.Geometry;
+using BH.oM.Geometry;
+
+namespace BH.Adapter.PowerPoint
+{
+    public partial class PowerPointAdapter : BHoMAdapter
+    {
+        /***************************************************/
+        /**** Interface Methods                         ****/
+        /***************************************************/
+
+        protected void IUpdateSlide(SlidePart slidePart, ISlideUpdate update)
+        {
+            if (update == null)
+                BH.Engine.Base.Compute.RecordError("No action was found for an update of type " + update.GetType().Name);
+            else
+                UpdateSlide(slidePart, update as dynamic);
+        }
+
+        /***************************************************/
+        /**** Fallback Methods                          ****/
+        /***************************************************/
+
+        private void UpdateSlide(SlidePart slidePart, ISlideUpdate update)
+        {
+            BH.Engine.Base.Compute.RecordError("No action was found for an update of type " + update.GetType().Name);
+        }
+
+
+        /***************************************************/
+        /**** Helper Methods                            ****/
+        /***************************************************/
+
+        private void CopyStream(Stream input, Stream output)
+        {
+            byte[] buffer = new byte[32768];
+            while (true)
+            {
+                int read = input.Read(buffer, 0, buffer.Length);
+                if (read <= 0)
+                    return;
+                output.Write(buffer, 0, read);
+            }
+        }
+
+        /***************************************************/
+    }
+}
+
+
